bump the lockfile, use the 'status' instead of the 'state' header and add alertmanager support  (#14)
* chore: bump lockfile

* fix: Rocket build

Rocket had breaking changes
[upstream](https://github.com/rwf2/Rocket/blob/master/CHANGELOG.md?plain=1#L327),
which are fixed in this commit

* chore: switch the state label to the status label

When looking at the documentation, there was a note about the 'state'
label being deprecated soon. The label 'status' is well-defined and
suits our purpose of indicating if we received a firing or resolved
message well

* priority: add alertmanager support

When testing this application with Prometheus [Alertmanager](https://prometheus.io/docs/alerting/latest/configuration/#webhook_config) an error occurred, that there was no title. Therefore, we set a default title to 'Alertmanager'.

* chore: add documentation about alertmanager support